refactor: move notional cells for claim supply rewards closer to their only usage
